+++ R/utils_addHelpTooltip.R
@@ -1,0 +1,13 @@
+#' This helper function add popover toolbars to all protocol elements
+#'
+#' @description
+#' Although this could be done directly in the protocol,
+#' it generally remains clear to add the help documentation directly here.
+#'
+#' @return Only adds to ui
+#' @noRd
+add_protocol_help <- function(session){
+  ns <- session$ns
+
+  # FIXME: Idea to outsource code here. Currently not working in app_server.R
+}