+++ dev/run_dev.R
@@ -1,5 +1,5 @@
 # Set options here
-options(golem.app.prod = F) # TRUE = production mode, FALSE = development mode
+options(golem.app.prod = T) # TRUE = production mode, FALSE = development mode
 
 # Comment this if you don't want the app to be served on a random port
 options(shiny.port = httpuv::randomPort())