+++ NEWS.md
@@ -1,3 +1,9 @@
+## Version 0.2
+
+* Added option to support different planning unit types for regular and irregular polygons.
+* Changed spatial grain of planning unit to provide input in case of a heterogeneous PU.
+* Added a few further text examples to elements, ultimately to be replaced by popovers #6
+
 ## Version 0.1
 
-* Initial Publication of the protocol and elicitation of expert feedback.
+* Initial Publication of the protocol.